Move Install/PoweredBy middleware to the grand.framework project
