+++ Grand.Framework/Infrastructure/DependencyRegistrar.cs
@@ -8,6 +8,7 @@ using Grand.Core.Http;
 using Grand.Core.Infrastructure;
 using Grand.Core.Infrastructure.DependencyManagement;
 using Grand.Core.Plugins;
+using Grand.Framework.Middleware;
 using Grand.Framework.Mvc.Routing;
 using Grand.Framework.TagHelpers;
 using Grand.Framework.Themes;
