+++ Grand.Framework/Infrastructure/Extensions/ApplicationBuilderExtensions.cs
@@ -4,6 +4,7 @@ using Grand.Core.Data;
 using Grand.Core.Domain;
 using Grand.Core.Http;
 using Grand.Core.Infrastructure;
+using Grand.Framework.Middleware;
 using Grand.Framework.Mvc.Routing;
 using Grand.Services.Authentication;
 using Grand.Services.Logging;
